# swagger.yaml
openapi: 3.0.0
info:
  title: TWITEE-API
  contact:
    name: MOA
    url: https://github.com/MOA-CODES
    email: moa.businessbread@gmail.com
  version: '1.0'
servers:
- url: https://twitee-api-mocha.vercel.app/api/v1
  variables: {}
paths:
...
  /twit:
    post:
      tags:
      - Twit
      summary: Make a Twit
      operationId: MakeaTwit
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/MakeaTwitRequest'
              - example:
                  post: Hello Twitee im Moa
            example:
              post: Hello Twitee im Moa
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Twit
      summary: Get My Twits
      operationId: GetMyTwits
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /twit/comment/{id}:
    parameters:
      - in: path
        name: id
        schema:
         type: string
        required: true
        description: The twit ID
  
    get:
      tags:
      - Twit
      summary: Get Twit Comments
      operationId: GetTwitComments
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /comment:
    get:
      tags:
      - Comment
      summary: Get My Comments
      operationId: GetMyComments
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []  
...
components:
  schemas:
...
    MakeaTwitRequest:
      title: MakeaTwitRequest
      required:
      - post
      type: object
      properties:
        post:
          type: string
      example:
        post: Hello Twitee im Moa
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...